requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
easyocr==1.7.0
opencv-python==4.8.1.78
pillow==10.2.0
numpy==1.24.3